make document-new a component
